requirements.txt
requests>=2.31.0
pyyaml>=6.0
dnspython>=2.4.0
schedule>=1.2.0
rich>=13.0.0
jinja2>=3.1.0
cryptography>=41.0.0
